refactor(filters): move filter definitions out of MenuBar into src/filters
Filters are an engine concept, not a menu concept — hosting them under
src/app/MenuBar/ was a "I added a Filter menu, where do these files go"
accident. Relocates all 19 filter declarations alongside curves/levels
in src/filters/, renames the index file to filter-registry.ts and the
type file to filter-types.ts, and updates the one importer in
filter-actions.ts.

Each filter is a one-shot GPU dispatch with a FilterDefinition — the
same structure as the adjustments already in src/filters/, so mixing
them there is the right locality. No behavior change.
