+++ src/filters/box-blur.ts
@@ -1,5 +1,5 @@
-import { filterBoxBlur } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterBoxBlur } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const boxBlur: FilterDefinition = {
   id: 'box-blur',
