+++ src/filters/chromatic-aberration.ts
@@ -1,5 +1,5 @@
-import { filterChromaticAberration } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterChromaticAberration } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const chromaticAberration: FilterDefinition = {
   id: 'chromatic-aberration',
