+++ src/filters/clouds.ts
@@ -1,5 +1,5 @@
-import { filterClouds } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterClouds } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const clouds: FilterDefinition = {
   id: 'clouds',
