+++ src/filters/halftone.ts
@@ -1,5 +1,5 @@
-import { filterHalftone } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterHalftone } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const halftone: FilterDefinition = {
   id: 'halftone',
