+++ src/filters/kaleidoscope.ts
@@ -1,5 +1,5 @@
-import { filterKaleidoscope } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterKaleidoscope } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const kaleidoscope: FilterDefinition = {
   id: 'kaleidoscope',
