+++ src/filters/motion-blur.ts
@@ -1,5 +1,5 @@
-import { filterMotionBlur } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterMotionBlur } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const motionBlur: FilterDefinition = {
   id: 'motion-blur',
