+++ src/filters/solarize.ts
@@ -1,5 +1,5 @@
-import { filterSolarize } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterSolarize } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const solarize: FilterDefinition = {
   id: 'solarize',
