+++ src/filters/unsharp-mask.ts
@@ -1,5 +1,5 @@
-import { filterUnsharpMask } from '../../../engine-wasm/wasm-bridge';
-import type { FilterDefinition } from './types';
+import { filterUnsharpMask } from '../engine-wasm/wasm-bridge';
+import type { FilterDefinition } from './filter-types';
 
 export const unsharpMask: FilterDefinition = {
   id: 'unsharp-mask',
